refactor(vjsc): remove unused helpers
`componentSourcePlugin` only served tests, so it moves under the plugin
tests as a helper. `collectFunctionDeclarations`, `findJsxElement`,
`prependBlockBody`, `InferParts`, and `InferPartProps` had no callers.
The traverse test locates its fixture element with `oxc-walker` directly.
