+++ packages/vjsc/src/ast/index.ts
@@ -9,7 +9,6 @@ export type {
   VariableDeclarator,
 } from '@oxc-project/types';
 export { walk } from 'oxc-walker';
-export { type BlockBody, prependBlockBody } from './edit';
 export {
   collectIdentifierNames,
   insertModuleImports,
@@ -25,4 +24,4 @@ export {
   type SourceText,
   sliceSource,
 } from './source';
-export { collectFunctionDeclarations, findJsxAttribute, findJsxElement, jsxNamePath } from './traverse';
+export { findJsxAttribute, jsxNamePath } from './traverse';
