+++ packages/vjsc/src/plugins/tests/component-target.test.ts
@@ -5,8 +5,8 @@ import { defineComponent, defineSchema } from '../../components/definition';
 import { defineComponentTarget } from '../../target/definition';
 import { Host, jsx } from '../../target/jsx-runtime';
 import { readComponentSource } from '../component-meta';
-import { componentSourcePlugin } from '../component-source';
 import { type ComponentTargetSelection, componentTargetPlugin } from '../component-target';
+import { componentSourcePlugin } from './helpers/component-source';
 
 const MODULE_ID = '\0fixture.tsx?target=react';
